chore: renamed index to types in themes
