+++ packages/graphiql/src/new-components/themes/types.ts
